feature/401-github-codeql producer
rename github code scanning to github codeql to better showcase what tool is used
