+++ components/producers/github-codeql/task.yaml
@@ -1,0 +1,46 @@
+---
+apiVersion: tekton.dev/v1beta1
+kind: Task
+metadata:
+  name: producer-github-codeql
+  labels:
+    v1.dracon.ocurity.com/component: producer
+    v1.dracon.ocurity.com/test-type: sast
+spec:
+  description: Retrieve a GitHub Code Scanning report from a GitHub repository.
+  params:
+  - name: producer-github-codeql-repository-owner
+    description: The owner of the repository to scan.
+    type: string
+  - name: producer-github-codeql-repository-name
+    description: The name of the repository to scan.
+    type: string
+  - name: producer-github-codeql-github-token
+    description: The GitHub token to use for scanning. Must have "Code scanning alerts" repository permissions (read).
+    type: string
+  volumes:
+    - name: scratch
+      emptyDir: {}
+  workspaces:
+    - name: output
+      description: The workspace containing the source-code to scan.
+  steps:
+  - name: produce-issues
+    imagePullPolicy: IfNotPresent
+    # image: '{{ default "ghcr.io/ocurity/dracon" .Values.image.registry }}/components/producers/github-codeql:{{ .Chart.AppVersion }}'
+    image: kind-registry:5000/components/producers/github-
+    command: ["/app/components/producers/github-codeql/github-codeql-parser"]
+    env:
+      - name: GITHUB_CLIENT_REQUEST_TIMEOUT
+        value: 5m
+      - name: GITHUB_CLIENT_LIST_PAGE_SIZE
+        value: "100"
+    args:
+    - "-in=/scratch/out.json"
+    - "-out=$(workspaces.output.path)/.dracon/producers/github-codeql.pb"
+    - "-github-token=$(params.producer-github-codeql-github-token)"
+    - "-repository-owner=$(params.producer-github-codeql-repository-owner)"
+    - "-repository-name=$(params.producer-github-codeql-repository-name)"
+    volumeMounts:
+    - mountPath: /scratch
+      name: scratch
